small fixes from looking at Jacques Basaldua's build with VisualC++
git-svn-id: svn://svn.code.sf.net/p/fuego/code/trunk@1109 2e953b5c-c64d-0410-be54-f773e93e544c

+++ go/GoNodeUtil.cpp
@@ -34,7 +34,8 @@ GoKomi GoNodeUtil::GetKomi(const SgNode* node)
         {
             try
             {
-                return GoKomi(node->GetRealProp(SG_PROP_KOMI));
+                return GoKomi(static_cast<float>
+                              (node->GetRealProp(SG_PROP_KOMI)));
             }
             catch (const GoKomi::InvalidKomi&)
             {
